dungeon cutscene split jirachi
