+++ data/data_8107224.s
@@ -12,26 +12,5 @@
 .string "pksdir0\0"
 .string "pksdir0\0"
 .string "pksdir0\0" @ 
-@ ??? Jirachi
-.string "pksdir0\0"
-
-.global gUnknown_81074FC
-gUnknown_81074FC: @ 81074FC
-.byte 0x4e, 0x4f, 0x50, 0x51, 0x4a, 0x49, 0x3e, 0x38
-
-@ ???
-.string "pksdir0\0" @ lugia
-@ ???
-.string "pksdir0\0" @kyorge
-@ ???
-.string "pksdir0\0" @ deoxys
-@ ???
-.string "pksdir0\0" @celebi
-@ ???
-.string "pksdir0\0" @unk
-@ ???
-.string "pksdir0\0" @ medicham
-@ ???
-.string "pksdir0\0" @ smeargle
 
 .align 2,0
