+++ ld_script.txt
@@ -289,6 +289,14 @@ SECTIONS {
         src/dungeon_cutscene_articuno.o(.text);
         src/dungeon_cutscene_empty_summit.o(.text);
         src/dungeon_cutscenes.o(.text);
+        src/dungeon_cutscene_jirachi.o(.text);
+        src/dungeon_cutscene_lugia.o(.text);
+        src/dungeon_cutscene_kyogre.o(.text);
+        src/dungeon_cutscene_deoxys.o(.text);
+        src/dungeon_cutscene_celebi.o(.text);
+        src/dungeon_cutscene_medicham.o(.text);
+        src/dungeon_cutscene_smeargle.o(.text);
+        src/dungeon_cutscene_no_one_here.o(.text);
         src/dungeon_portrait_placement.o(.text);
         src/pokemon.o(.text);
         src/pokemon_3.o(.text);
@@ -636,6 +644,14 @@ SECTIONS {
         src/dungeon_cutscene_articuno.o(.rodata);
         src/dungeon_cutscene_empty_summit.o(.rodata);
         data/data_8107224.o(.rodata);
+        src/dungeon_cutscene_jirachi.o(.rodata);
+        src/dungeon_cutscene_lugia.o(.rodata);
+        src/dungeon_cutscene_kyogre.o(.rodata);
+        src/dungeon_cutscene_deoxys.o(.rodata);
+        src/dungeon_cutscene_celebi.o(.rodata);
+        src/dungeon_cutscene_medicham.o(.rodata);
+        src/dungeon_cutscene_smeargle.o(.rodata);
+        src/dungeon_cutscene_no_one_here.o(.rodata);
         src/dungeon_portrait_placement.o(.rodata);
         src/pokemon.o(.rodata);
         data/data_81076E4.o(.rodata);
