+++ src/dungeon_cutscene_no_one_here.c
@@ -1,0 +1,13 @@
+#include "global.h"
+#include "globaldata.h"
+#include "code_80869E4.h"
+#include "dungeon_boss_dialogue.h"
+
+void sub_808CD9C(void)
+{
+    sub_8086448();
+    // There appears to be no one here.
+    // It's impossible to go any further
+    sub_80866C4(&gUnknown_810697C);
+}
+
